requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
pdfservices-sdk==4.0.0
pathlib==1.0.1
Pillow==10.1.0